requirements.txt
Flask
Flask-SQLAlchemy
pandas
psycopg2-binary
openpyxl
Flask-RESTful
joblib
dotenv
python-dotenv
Flask-Migrate
